cannot negate size_t. Only a problem on 32-bit OSes

+++ moses/DummyScoreProducers.cpp
@@ -112,7 +112,7 @@ void WordPenaltyProducer::Evaluate(const TargetPhrase &targetPhrase
                     , ScoreComponentCollection &scoreBreakdown
                     , ScoreComponentCollection &estimatedFutureScore) const
 {
-  float score = - targetPhrase.GetNumTerminals();
+  float score = - (float) targetPhrase.GetNumTerminals();
   scoreBreakdown.Assign(this, score);
 }
